Move shared code and tests for that code
