+++ cmd/common_test.go
@@ -1,0 +1,45 @@
+package main
+
+import (
+	"io/ioutil"
+	"os"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func newConfigFile(t *testing.T, b []byte) string {
+	f, err := ioutil.TempFile("", "aws-config")
+	if err != nil {
+		t.Fatal(err)
+	}
+	if err := ioutil.WriteFile(f.Name(), b, 0600); err != nil {
+		t.Fatal(err)
+	}
+	return f.Name()
+}
+
+func TestGenerateQrCode(t *testing.T) {
+	tempFile, err := ioutil.TempFile("", "temp-qr.*.png")
+	assert.NoError(t, err)
+	defer func() {
+		errRemove := os.Remove(tempFile.Name())
+		assert.NoError(t, errRemove)
+	}()
+
+	err = generateQrCode("otpauth://totp/super@top?secret=secret", tempFile)
+	assert.NoError(t, err)
+}
+
+func TestGetPartition(t *testing.T) {
+	commPartition, err := getPartition("us-west-2")
+	assert.Equal(t, commPartition, "aws")
+	assert.NoError(t, err)
+
+	govPartition, err := getPartition("us-gov-west-1")
+	assert.Equal(t, govPartition, "aws-us-gov")
+	assert.NoError(t, err)
+
+	_, err = getPartition("aws-under-the-sea")
+	assert.Error(t, err)
+}
