+++ cmd/setup_test.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"io/ioutil"
 	"log"
 	"os"
 	"testing"
@@ -9,17 +8,6 @@ import (
 	"github.com/99designs/aws-vault/vault"
 	"github.com/stretchr/testify/assert"
 )
-
-func newConfigFile(t *testing.T, b []byte) string {
-	f, err := ioutil.TempFile("", "aws-config")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := ioutil.WriteFile(f.Name(), b, 0600); err != nil {
-		t.Fatal(err)
-	}
-	return f.Name()
-}
 
 func TestUpdateAWSConfigFile(t *testing.T) {
 
@@ -84,29 +72,4 @@ output=json
 	assert.Equal(t, testSection.RoleARN, "arn:aws:iam::123456789012:role/test-role")
 	assert.Equal(t, testSection.Region, "us-west-2")
 	// assert.Equal(t, testBaseSection.Output, "json")
-}
-
-func TestGenerateQrCode(t *testing.T) {
-	tempFile, err := ioutil.TempFile("", "temp-qr.*.png")
-	assert.NoError(t, err)
-	defer func() {
-		errRemove := os.Remove(tempFile.Name())
-		assert.NoError(t, errRemove)
-	}()
-
-	err = generateQrCode("otpauth://totp/super@top?secret=secret", tempFile)
-	assert.NoError(t, err)
-}
-
-func TestGetPartition(t *testing.T) {
-	commPartition, err := getPartition("us-west-2")
-	assert.Equal(t, commPartition, "aws")
-	assert.NoError(t, err)
-
-	govPartition, err := getPartition("us-gov-west-1")
-	assert.Equal(t, govPartition, "aws-us-gov")
-	assert.NoError(t, err)
-
-	_, err = getPartition("aws-under-the-sea")
-	assert.Error(t, err)
 }
